services: move fetch functions
